requirements.txt
aiohttp==0.22.5
amqp==1.4.9
anyjson==0.3.3
asyncio==3.4.3
billiard==3.3.0.23
cchardet==1.0.0
celery==3.1.23
certifi==2016.8.8
chardet==2.3.0
click==6.6
future==0.15.2
kombu==3.0.35
multidict==1.2.2
psycopg2==2.6.2
python-telegram-bot==5.0.0
pytz==2016.6.1
redis==2.10.5
rq==0.6.0
SQLAlchemy==1.0.14
urllib3==1.16
